Use common structure for parsing different kinds of payload, fixed reconnect logic

+++ pe-terminal.go
@@ -120,10 +120,10 @@ func main() {
 			}
 		}
 	}
-	tunnel.OnResize = func(sessionID string, width uint16, height uint16) {
+	tunnel.OnResize = func(sessionID string, width int64, height int64) {
 		if _, ok := sessionsMap[sessionID]; ok {
 			log.Printf("Resize terminal w: %v, h: %v from %s\n", width, height, sessionID)
-			sessionsMap[sessionID].Resize(width, height)
+			sessionsMap[sessionID].Resize(uint16(width), uint16(height))
 		}
 	}
 	tunnel.OnError = func(err error) {
